fix: restructure settings/lib/ to api/hooks/utils/types + fix type errors
- Move 13 files from settings/lib/ to api/, hooks/, utils/, types/
  to match the destination repo's no-lib-in-domains convention
- Track web-search-provider-catalog.gen.ts (static provider catalog)
- Remove unused WEB_SEARCH_PROVIDER_CATALOG array and types
- Fix TS2769 Object.entries overload in share-feedback-modal.tsx
- Update 25 import paths across settings domain
