+++ apps/web/src/domains/settings/components/create-schedule-modal.tsx
@@ -6,7 +6,7 @@ import { Modal } from "@vellum/design-library/components/modal";
 import {
   createSchedule,
   type CreateSchedulePayload,
-} from "@/domains/settings/lib/schedules/api.js";
+} from "@/domains/settings/api/schedules.js";
 
 // ---------------------------------------------------------------------------
 // Cron presets — cover the most common cases without forcing users to learn
