+++ apps/web/src/domains/settings/components/trust-rules/trust-rules-modal.tsx
@@ -8,8 +8,8 @@ import { ConfirmDialog } from "@vellum/design-library/components/confirm-dialog"
 import { Toggle } from "@vellum/design-library/components/toggle";
 import { Notice } from "@vellum/design-library/components/notice";
 import { Tag, type TagTone } from "@vellum/design-library/components/tag";
-import { deleteTrustRule, fetchTrustRules } from "@/domains/settings/lib/trust-rules/api.js";
-import type { TrustRuleItem, TrustRuleRisk } from "@/domains/settings/lib/trust-rules/types.js";
+import { deleteTrustRule, fetchTrustRules } from "@/domains/settings/api/trust-rules.js";
+import type { TrustRuleItem, TrustRuleRisk } from "@/domains/settings/types/trust-rules.js";
 
 import { TrustRuleFormModal } from "@/domains/settings/components/trust-rules/trust-rule-form-modal.js";
