+++ apps/web/src/domains/settings/types/schedules.ts
