+++ apps/web/src/domains/settings/types/trust-rules.ts
