+++ apps/web/src/domains/settings/utils/notification-utils.ts
