+++ apps/web/src/domains/settings/utils/theme-preferences.ts
